feat: move file generation queue to worker
